requirements.txt
Flask
Pillow
lunar-python
requests
pytz
beautifulsoup4
qrcode